refactor(`-lesson-plan`): unify code related to lesson plan document organization and formation
