+++ code/drasil-lesson-plan/lib/Drasil/LessonPlan.hs
@@ -1,7 +1,7 @@
 module Drasil.LessonPlan (
-  module Drasil.LessonPlan.Core,
+  module Drasil.LessonPlan.Document,
   module Drasil.LessonPlan.Renderer
 ) where
 
-import Drasil.LessonPlan.Core
+import Drasil.LessonPlan.Document
 import Drasil.LessonPlan.Renderer
