+++ code/drasil-lesson-plan/lib/Drasil/LessonPlan/Document.hs
@@ -1,4 +1,10 @@
-module Drasil.LessonPlan.Sections (
+module Drasil.LessonPlan.Document (
+  -- * Lesson Plan
+  LsnDesc, LsnChapter(..),
+  -- ** Chapter Types
+  Intro(..), LearnObj(..), Review(..), CaseProb(..),
+  Example(..), Smmry(..), Apndx(..),
+  -- ** Constructors
   intro, learnObj, caseProb, example, appendix, review, reference, summary
 ) where
 
@@ -7,6 +13,42 @@ import Language.Drasil.Document
 
 import qualified Drasil.Metadata.Documentation as Doc (caseProb, introduction,
   learnObj, review, summary, example, appendix, reference)
+
+-- * Lesson Chapter Types
+
+type LsnDesc = [LsnChapter]
+
+data LsnChapter = Intro Intro
+                | LearnObj LearnObj
+                | Review Review
+                | CaseProb CaseProb
+                | Example Example
+                | Smmry Smmry
+                | BibSec
+                | Apndx Apndx
+
+-- TODO: Work on detail structure of Lesson Plan
+
+-- ** Introduction
+newtype Intro = IntrodProg [Contents]
+
+-- ** Learning Objectives
+newtype LearnObj = LrnObjProg [Contents]
+
+-- ** Review Chapter
+data Review = ReviewProg [Contents] [Section]
+
+-- ** A Case Problem
+data CaseProb = CaseProbProg [Contents] [Section]
+
+-- ** Examples of the lesson
+newtype Example = ExampleProg [Contents]
+
+-- ** Summary
+newtype Smmry = SmmryProg [Contents]
+
+-- ** Appendix
+newtype Apndx = ApndxProg [Contents]
 
 -- * Section Constructors
